add cart and language feature
The site now has available a cart where you can add multiple items from the DuckStore tab. The user cart can be accessed by clicking the shopping bag on top of the page next to the DuckTabs logo. Everytime you add an item a modal warns you that you added the item to the cart. Also now you can change the language of the web site simply by choosing one of the options of the dropdown selector next to the shopping bag.